refactor(types): restructure and modularize type definitions
Consolidate related type definitions into dedicated subdirectories and
rename generic type files to improve codebase organization and domain
clarity.

- Relocate character-specific interfaces to `src/types/character/`
- Move world editor definitions to `src/types/worldeditor/`
- Rename the generic `types.ts` to `worldbook.ts`
- Update all import paths across components, composables, and utilities
- Remove obsolete top-level type files
