+++ src/components/cardManager/components/CharacterCardHome.vue
@@ -177,7 +177,7 @@
 <script setup lang="ts">
 import BrowserFilePicker from '@/components/ui/common/BrowserFilePicker.vue';
 import { formatDate, now, toDateSafe } from '@/utils/datetime';
-import type { CharacterCardCollection } from '@/types/character-card-collection';
+import type { CharacterCardCollection } from '@/types/character/character-card-collection';
 import { Delete, Download, FolderOpened, MoreFilled, Plus, Search } from '@element-plus/icons-vue';
 import { Icon } from '@iconify/vue';
 import {
