+++ src/components/cardManager/components/WorldBookSelectorDialog.vue
@@ -140,7 +140,7 @@ import { ref, computed, watch } from 'vue';
 import { ElDialog, ElScrollbar, ElButton, ElTag, ElEmpty, ElCollapse, ElCollapseItem } from 'element-plus';
 import { Icon } from '@iconify/vue';
 import { worldBookService } from '@/database/worldBookService';
-import type { WorldBook } from '@/types/types';
+import type { WorldBook } from '@/types/worldbook';
 import { useRouter } from 'vue-router';
 
 interface Props {
