+++ src/components/cardManager/main/AdvancedInfoPanel.vue
@@ -105,7 +105,7 @@
 
 <script setup lang="ts">
 import { ElFormItem, ElInput, ElRow, ElCol } from 'element-plus';
-import type { CharacterCardV3 } from '@/types/character-card-v3';
+import type { CharacterCardV3 } from '@/types/character/character-card-v3';
 
 defineProps<{
   character: CharacterCardV3;
