+++ src/components/cardManager/panel/CardWorldBookPanel.vue
@@ -310,9 +310,9 @@ import WorldBookSelectorDialog from '@/components/cardManager/components/WorldBo
 import { useWorldBookDragDrop } from '@/composables/worldbook/useWorldBookDragDrop';
 import { useWorldBookEntry } from '@/composables/worldbook/useWorldBookEntry';
 import { worldBookService } from '@/database/worldBookService';
-import type { CharacterBook } from '@/types/character-book';
-import type { CharacterCardV3 } from '@/types/character-card-v3';
-import type { WorldBook, WorldBookCollection, WorldBookEntry } from '@/types/types';
+import type { CharacterBook } from '@/types/character/character-book';
+import type { CharacterCardV3 } from '@/types/character/character-card-v3';
+import type { WorldBook, WorldBookCollection, WorldBookEntry } from '@/types/worldbook';
 import { convertCharacterBookToWorldBook, convertWorldBookToCharacterBook } from '@/utils/worldBookConverter';
 import ConfirmDialog from '../components/ConfirmDialog.vue';
