+++ src/composables/characterCard/useCharacterCardCollection.ts
@@ -2,8 +2,8 @@ import { ref, computed, onMounted } from 'vue';
 import { ElMessage, ElMessageBox } from 'element-plus';
 import { v4 as uuidv4 } from 'uuid';
 import { read as readPngMetadata } from '@/utils/pngCardMetadata';
-import type { CharacterCardV3 } from '@/types/character-card-v3';
-import type { CharacterCardCollection, CharacterCardItem } from '@/types/character-card-collection';
+import type { CharacterCardV3 } from '@/types/character/character-card-v3';
+import type { CharacterCardCollection, CharacterCardItem } from '@/types/character/character-card-collection';
 import { characterCardService, type StoredCharacterCard } from '@/database/characterCardService';
 import { nowIso } from '@/utils/datetime';
 import { saveFile } from '@/utils/fileSave';
