+++ src/types/character/character-book.ts
