+++ src/types/character/character-card-v3.ts
