+++ src/types/character/character.ts
