+++ src/utils/worldeditor/graphGeometry.ts
@@ -1,4 +1,4 @@
-import type { EnhancedLandmark } from '@/types/world-editor';
+import type { EnhancedLandmark } from '@/types/worldeditor/world-editor';
 
 type HandleSide = 'left' | 'right' | 'top' | 'bottom';
